Moved to stateless persistence implementation
Until now, the persistence class held the state of the current active bundle. It was used to trigger file deletion from the outside after a successful call to the server. This wasn't ideal because this can cause a variety of issues. Now, as soon as we have successfully deserialized a bundle, the corresponding file is deleted immediately, which means that we have to store the bundle again in case of a failed request